// src/uiLibrary/interfaces/properties/IPropMargin.ts
export default interface IPropsMargin {
  margin?: number;
  margin1?: boolean;
  margin2?: boolean;
  margin3?: boolean;
  margin4?: boolean;
  margin5?: boolean;
  margin6?: boolean;
  margin7?: boolean;
  margin8?: boolean;
  marginBottom?: number;
  marginBottom1?: boolean;
  marginBottom2?: boolean;
  marginBottom3?: boolean;
  marginBottom4?: boolean;
  marginBottom5?: boolean;
  marginBottom6?: boolean;
  marginBottom7?: boolean;
  marginBottom8?: boolean;
  marginLeft?: number;
  marginLeft1?: boolean;
  marginLeft2?: boolean;
  marginLeft3?: boolean;
  marginLeft4?: boolean;
  marginLeft5?: boolean;
  marginLeft6?: boolean;
  marginLeft7?: boolean;
  marginLeft8?: boolean;
  marginRight?: number;
  marginRight1?: boolean;
  marginRight2?: boolean;
  marginRight3?: boolean;
  marginRight4?: boolean;
  marginRight5?: boolean;
  marginRight6?: boolean;
  marginRight7?: boolean;
  marginRight8?: boolean;
  marginTop?: number;
  marginTop1?: boolean;
  marginTop2?: boolean;
  marginTop3?: boolean;
  marginTop4?: boolean;
  marginTop5?: boolean;
  marginTop6?: boolean;
  marginTop7?: boolean;
  marginTop8?: boolean;
}